Finish multi-repo navigation and remove GOVREADY env var
Refactor routes.py to correctly see multiple repositories
and navigate between multiple repositories indicated in
local specific `.hypergrc_repos` simple text file listing repos.

The correct repo's .govready file is determined dynamically
by "organization" and "project" combination.

There is no longer a need to set an GOVREADY environment
variable.

This navigation is still considered safe, because it is
restricted to a single workstation and the files that are on
that workstation.