Hide loading gif on browser back button
Hacky solution to make sure the loading gif
does not show up when clicking browser back button.
Because loading gif is a simple hide/show onclicking a link
the loading gif would normally still be displayed when
clicking browser back button as it was last state of
previous page. This change corrects.