Custom background, bg-image, css read from project repos
Sure, it is a small thing with no impact on system compliance,
but it can make the hyperGRC nicer to use by allowing reading
custom css files from a project file. When we load the project,
hyperGRC checks to see if a custom repo.css file is present
in the project's indicated `_extensions` subdirectories and adds
the file path to the project data if found. The `base.html` page
will then render the extra css reference if present.