adding controls (ec499070070ed86ff1d86b487a75aee94ec33853) broke updating the UI after controls were edited
We have a client-side object holding information about each control displayed, and that object is used to initialize the control editor and to update the UI after an edit. In ec499070070ed86ff1d86b487a75aee94ec33853, I accidentally left out the part that updates that object after an edit is made. Editing the control for a second time would show the original control content, and the tooltip on control items was updated with the orginal control content instead of the new content.

Rather than updating the client-side object simply on the client side, the server now returns JSON data when controls are updated (via AJAX) with the control's updated data, and then we update the client-side object holding the control data from that. This ensures the client has exactly the same information that the server has.

clean_text is also updated --- it was removing blank lines in narratives because its regex for whitespace-at-the-end-of-a-line was too inclusive and also pulled in consecutive newlines.