Fix control_family sort crash mix'n known/unknown control families
Sometimes unknown control families show up and do not receive an integer assignment that known control_families do for search purposes. This led to Python trying to compare integers and  strings in sort and crashing program. To solve problem, we turned our sort key into a tuple to mix integer and strings and now we can deal with unknown control families in the data.